Modularized in multiple components
